Tier 2.3: Consolidate test helpers into support directory
Reorganized test infrastructure into centralized support/ directory for better
discoverability and organization.

Structure changes:
- Created src/__tests__/support/ directory
- Moved fixtures/ → support/fixtures/
- Moved helpers/ → support/helpers/
- Moved mocks/ → support/mocks/
- Moved unit/services/__helpers__/ → support/helpers/service-mocks/
- Moved integration/__helpers__/ → support/helpers/integration-utils/

Files moved: 17 files
- 7 fixture files (database, repository, static-fixtures, swarm, user, workspace, index)
- 7 helper files (api-assertions, auth, database-assertions, ids, request-builders, transactions, index)
- 3 mock files (env, prisma, index)

Import updates: 40+ imports updated across:
- 15 integration test files
- 2 unit test files
- 2 setup files
- Test helper documentation

New index files created:
- support/helpers/service-mocks/index.ts
- support/helpers/integration-utils/index.ts

All tests passing:
- Unit: 1,270 tests | 63 files
- Integration: 163 tests | 15 files
