+++ src/__tests__/integration/api/workspace-member-roles.test.ts
@@ -4,8 +4,8 @@ import { PATCH } from "@/app/api/workspaces/[slug]/members/[userId]/route";
 import { WorkspaceRole } from "@prisma/client";
 import { AssignableMemberRoles } from "@/lib/auth/roles";
 import { db } from "@/lib/db";
-import { createTestWorkspaceScenario } from "@/__tests__/fixtures/workspace";
-import { createTestUser } from "@/__tests__/fixtures/user";
+import { createTestWorkspaceScenario } from "@/__tests__/support/fixtures/workspace";
+import { createTestUser } from "@/__tests__/support/fixtures/user";
 import {
   createAuthenticatedSession,
   mockUnauthenticatedSession,
@@ -16,7 +16,7 @@ import {
   createPostRequest,
   createPatchRequest,
   getMockedSession,
-} from "@/__tests__/helpers";
+} from "@/__tests__/support/helpers";
 
 // Mock GitHub API calls for addWorkspaceMember (external service)
 vi.mock("@/services/github", () => ({
