+++ src/__tests__/support/fixtures/static-fixtures.ts
