+++ src/__tests__/utils/test-helpers.ts
@@ -1,3 +1,3 @@
-// Deprecated: prefer importing from `@/__tests__/fixtures` or `@/__tests__/assertions`.
-export * from "../fixtures";
+// Deprecated: prefer importing from `@/__tests__/support/fixtures` or `@/__tests__/assertions`.
+export * from "../support/fixtures";
 export * from "../assertions";
